Update FieldLocalisation and FIeldLineDetector

+++ module/localisation/FieldLocalisation/src/FieldLocalisation.hpp
@@ -20,8 +20,8 @@ namespace module::localisation {
 
     // Particle struct
     struct Particle {
-        Eigen::Vector3f state = Eigen::Vector3f::Zero();  // (x, y, theta) of world space in field space
-        float weight          = 1.0;
+        Eigen::Vector3d state = Eigen::Vector3d::Zero();  // (x, y, theta) of world space in field space
+        double weight         = 1.0;
     };
 
     class FieldLocalisation : public NUClear::Reactor {
@@ -29,37 +29,37 @@ namespace module::localisation {
         /// @brief Stores configuration values
         struct Config {
             /// @brief Size of the grid cells in the occupancy grid [m]
-            float grid_size = 0.0;
+            double grid_size = 0.0;
             /// @brief Number of particles to use in the particle filter
             int n_particles = 0;
             /// @brief Uncertainty in the process model
-            Eigen::Matrix3f process_noise = Eigen::Matrix3f::Zero();
+            Eigen::Matrix3d process_noise = Eigen::Matrix3d::Zero();
             /// @brief Uncertainty in the measurement model
-            float measurement_noise = 0;
+            double measurement_noise = 0;
             /// @brief Maximum distance a field line can be from a particle to be considered an observation [m]
-            float max_range = 0;
+            double max_range = 0;
             /// @brief Initial state (x,y,theta) of the robot, saved for resetting
-            std::vector<Eigen::Vector3f> initial_state{};
+            std::vector<Eigen::Vector3d> initial_state{};
             /// @brief Initial covariance matrix of the robot's state, saved for resetting
-            Eigen::Matrix3f initial_covariance = Eigen::Matrix3f::Identity();
+            Eigen::Matrix3d initial_covariance = Eigen::Matrix3d::Identity();
             /// @brief Bool to enable/disable saving the generated map as a csv file
             bool save_map = false;
             /// @brief Minimum number of field line points for a measurement update
             size_t min_observations = 0;
             /// @brief Penalty factor for observations being outside map
-            float outside_map_penalty_factor = 0.0;
+            double outside_map_penalty_factor = 0.0;
         } cfg;
 
         NUClear::clock::time_point last_time_update_time;
 
         /// @brief Occupancy grid map of the field lines
-        OccupancyMap<float> fieldline_map;
+        OccupancyMap<double> fieldline_map;
 
         /// @brief State (x,y,theta) of the robot
-        Eigen::Vector3f state = Eigen::Vector3f::Zero();
+        Eigen::Vector3d state = Eigen::Vector3d::Zero();
 
         /// @brief Covariance matrix of the robot's state
-        Eigen::Matrix3f covariance = Eigen::Matrix3f::Identity();
+        Eigen::Matrix3d covariance = Eigen::Matrix3d::Identity();
 
         /// @brief Status of if the robot is falling
         bool falling = false;
@@ -76,21 +76,21 @@ namespace module::localisation {
         /// @param particle The state of the particle (x,y,theta)
         /// @param rPWw The field point (x, y) in world space {w} [m]
         /// @return The observation location (x, y) in the map
-        Eigen::Vector2i position_in_map(const Eigen::Vector3f particle, const Eigen::Vector3f rPWw);
+        Eigen::Vector2i position_in_map(const Eigen::Vector3d particle, const Eigen::Vector3d rPWw);
 
         /// @brief Get the weight of a particle given a set of observations
         /// @param particle The state of the particle (x,y,theta)
         /// @param observations The observations (x, y) in the robot's coordinate frame [m]
         /// @return The weight of the particle
-        float calculate_weight(const Eigen::Vector3f particle, const std::vector<Eigen::Vector3f>& observations);
+        double calculate_weight(const Eigen::Vector3d particle, const std::vector<Eigen::Vector3d>& observations);
 
         /// @brief Get the current mean (state) of the robot
         // @return The current mean (state) of the robot
-        Eigen::Vector3f compute_mean();
+        Eigen::Vector3d compute_mean();
 
         /// @brief Get the current covariance matrix of the robot's state
         // @return The current covariance matrix of the robot's state
-        Eigen::Matrix3f compute_covariance();
+        Eigen::Matrix3d compute_covariance();
 
         /// @brief Perform a time update on the particles
         /// @param walk_command The walk command (dx, dy, dtheta)
